+++ scripts/influxdb-onboarding.sh
@@ -23,7 +23,7 @@
 
 set -e
 
-echo "Wait to start InfluxDB 2.0"
+echo "Wait to start InfluxDB 2.x"
 wget -S --spider --tries=20 --retry-connrefused --waitretry=5 http://localhost:8086/metrics
 
 echo