+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 2.7.0 [unreleased]
 
+### Features
+1. [#119](https://github.com/influxdata/influxdb-client-php/pull/119): Add `InvocableScriptsApi` to create, update, list, delete and invoke scripts by seamless way
+
 ### CI
 1. [#118](https://github.com/influxdata/influxdb-client-php/pull/118): Use new Codecov uploader for reporting code coverage